requirements.txt
pathway[xpack-llm]
python-dotenv
groq
pdf2image
pypdf
unstructured[pdf]